Extract ProductType struct to the separate file
